Continue refactor by decoupling declarative from executive data.
Set up Measurement architecture. Now there are *Declaration classes that
capture the declarative side of data, and *Record classes that capture the
pieces of data as seen during execution.

Add a first-pass at unit code replacement in util/units.py.

The hello_world example "runs", but measurement values are not actually set
as SetValue is currently a stub and needs to be figured out and filled in.