Remove capture_test_source flag
After using it internally, I realized this flag can't work the way it's used right now. By definition, the phases are defined before the test is created or executed, but we parse flags right before we `Execute` a test. Which means that the `capture_test_source` flag was misleading: it was always used before its value was known.