Add an explicit Equals validator for non-basic types
This allows equality checks to be done for measurements that aren't strings or integers. My use-case is a list of button presses that I want to verify are done in the right order. I could create N measurements and fill them all in, or just 1 measurement and put a list in directly.