Remove a Condition.wait(None) from user_input
It turns out, the kryptonite to KillableThread is a thread waiting on a Condition with a timeout of None. If the current phase (such as a test_start trigger, or any user phase) is waiting on a Condition because of user_input.UserInput, then it'll never exit until the prompt is answered.