Don't wait for the multicast listener to stop when stopping the StationApi server
The multicast listener can wait up to util.multicast.MulticastListener.LISTEN_TIMEOUT_S (60) seconds before exiting, which is rather annoying when you want to stop the StationApi server. This passes a timeout of 0 seconds (aka, don't wait) for how long we want to wait for it to exit when we're being stopped too.