Do not pickle the TestRecord in the station API
The TestRecord may include things that the tester binary knows about,
such as numpy arrays (as measurement data, for instance), but that the
frontend doesn't (it may be built in a different virtualenv). In these
scenarios, the frontend cannot unpickle a TestRecord, so we first
convert to basic data types in order to avoid such problems.