Allow phases to return SKIP (#587)
Allow phases to return SKIP after they have started in case they need
some additional data to determine if they should actually run.

This functions differently than run_if in that if run_if returns false,
the phase is not logged.  If the phase returns SKIP, it will be logged.