Handle dimensioned measurement validation errors (#770)
Exceptions in dimensioned measurement validations occur in the Test
Executor thread, which measn they are not caught by the normal flow of
tests.  When those occur, the test will end and will be reported as a
PASS.

This commit changes that.  Now, an exception during a dimensioned
validation measurement validator function will set the phase result to
the raise exception if the phase result is not terminal.  The existing
result will be left in place if it is terminal; in this case, the
validator exception is logged with its stack dump to ensure that there
is still a record.