Use custom logger subclass for test records. (#862)
The test-specific logger is now a custom subclass that overrides the
getChild method to avoid the Python logging module's log hierarchy.
Subloggers must now call the parent's getChild method to construct the
appropriate logger.

In addition, this change moves creating the phase-specific logger to
TestState's running_phase_context.  This ensures that the logger is only
created once for a particular phase, rather than creating multiples in
scenarios where phases recursively call other phases. To make this
clearer, the TestState.logger has been renamed to
TestState.state_logger; TestState.logger is now a property that returns
the running phase state's logger instance or exceptions if it is not
set.

Lastly, fix flakiness in user_input that was causing unit tests to
intermittently failing.