station_server: provide history handlers
While the history handlers do need to match what has been saved on disk,
the project provides MfgEvent protobuffer writers out of the box, and
no-others, so at least let the out of box experience function as
expected, even if you would want to change this in your own
implementations.

To enable the (built in) writers, something like this is required in
your station server.

```

if __name__ == '__main__':
	openhtf.util.conf.load(station_server_port='4444')

	interface = mfg_inspector.MfgInspector()
	interface.set_converter(mfg_event_from_test_record)

	with station_server.StationServer(history_path="somepath") as server:
		while 1:
			test = .... #your tests here
			test.add_output_callbacks(server.publish_final_state)
			# explicitly match hardcoded pattern in HistoryListHandler
			test.add_output_callbacks(interface.save_to_disk("somepath/mfg_event_{dut_id}_{start_time_millis}.pb"))
			test.execute(test_start=user_input.prompt_for_test_start())
```

Signed-off-by: Karl Palsson <karlp@etactica.com>