Change multicast error handling so that it ignores unretriable errors.
The multicast listener attempts to add local and default interfaces to multicast, but on machines that only have external IPv6 interfaces, it won't find the default interface, and was getting stuck retrying.

This changes error handling to move on from interfaces that throw unretriable errors, determined by the errno of the OSError. This doesn't fix multicast for IPv6, but it means that the localhost multicast works regardless.

PiperOrigin-RevId: 786876667